+++ 99.wc_object/wc_object.rb
@@ -1,0 +1,7 @@
+#!/usr/bin/env ruby
+
+# frozen_string_literal: true
+
+require_relative 'wc'
+
+Wc.new(ARGV, ARGF).display
